Remove queueable handlers
